refactored meetings and removed scraping.kt
